services/poker-engine/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-socketio==5.11.2
supabase==2.4.2
python-dotenv==1.0.1
treys==0.1.8